// src/app/models/inventoryModel.ts
export class InventoryData {    
    flightNumber: string = '';
    airlineNo: string = '';
    fromPlace: string = '';
    toPlace: string = '';
    startDateTime: string = '';
    endDateTime: string = '';
    scheduleDays: string = '';
    instrumentUsed: string = '';
    businessClassSeat: number = 0;
    nonBusinessClassSeat: number = 0;
    ticketCost: number = 0;
    noOfRows: number = 0;
    meal: string = '';
}